writing test and apply multiple schemas to testcontainer postgres

+++ src/application/usecase/post/delete_post.ts
@@ -1,6 +1,10 @@
 import type { PostRepo } from "#application/repo/post_repo.ts";
+import { BaseError, BaseErrorEnum } from "#domain/error.ts";
 
-type Input = { id: string };
+type Input = {
+  id: string;
+  currentAuthorId: string;
+};
 
 type Output = void;
 
@@ -12,6 +16,21 @@ export class DeletePost {
   }
 
   async handle(input: Input): Promise<Output> {
+    const item = await this.#postRepo.getById(input.id);
+    if (item === null) {
+      return;
+    }
+
+    if (item.authorId !== input.currentAuthorId) {
+      throw new DeleteNotAuthorizedError();
+    }
+
     await this.#postRepo.delete(input.id);
+  }
+}
+
+export class DeleteNotAuthorizedError extends BaseError {
+  constructor() {
+    super(BaseErrorEnum.authorization, "Not Authorized");
   }
 }
